Trait peepmatic_automata::dot::DotFmt

pub trait DotFmt<TAlphabet, TState, TOutput> {
    fn fmt_transition(
        &self, 
        w: &mut impl Write, 
        from: Option<&TState>, 
        input: &TAlphabet, 
        to: Option<&TState>
    ) -> Result<()>;

    fn fmt_state(&self, w: &mut impl Write, state: &TState) -> Result<()>;

    fn fmt_output(&self, w: &mut impl Write, output: &TOutput) -> Result<()>;
}

Format the user-provided bits of an Automaton for Graphviz Dot output.

There are two provided implementations of DotFmt:

DebugDotFmt – format each type parameter with its std::fmt::Debug implementation.

DisplayDotFmt – format each type parameter with its std::fmt::Display implementation.

You can also implement this trait yourself if your type parameters don’t implement Debug or Display, or if you want to format them in some other way.

Required methods

fn fmt_transition(
&self,
w: &mut impl Write,
from: Option<&TState>,
input: &TAlphabet,
to: Option<&TState>
) -> Result<()>

Format a transition edge: from ---input---> to.

This will be inside an HTML label, so you may use balanced HTML tags.

fn fmt_state(&self, w: &mut impl Write, state: &TState) -> Result<()>

Format the custom data associated with a state.

This will be inside an HTML label, so you may use balanced HTML tags.

fn fmt_output(&self, w: &mut impl Write, output: &TOutput) -> Result<()>

Format a transition’s output or the final output of a final state.

This will be inside an HTML label, so you may use balanced HTML tags.